[Related to: #72} Migration to Java 21 LTS and bloxk XSS
