+++ jamix-business/pom.xml
@@ -27,7 +27,7 @@
 		<url/>
 	</scm>
 	<properties>
-		<java.version>17</java.version>
+		<java.version>21</java.version>
 	</properties>
 	<dependencies>
 		<dependency>
@@ -61,6 +61,12 @@
     		<groupId>com.auth0</groupId>
     		<artifactId>java-jwt</artifactId>
     		<version>4.4.0</version>
+		</dependency>
+		<!-- https://mvnrepository.com/artifact/com.googlecode.owasp-java-html-sanitizer/owasp-java-html-sanitizer -->
+		<dependency>
+    		<groupId>com.googlecode.owasp-java-html-sanitizer</groupId>
+    		<artifactId>owasp-java-html-sanitizer</artifactId>
+    		<version>20240325.1</version>
 		</dependency>
 	</dependencies>
