+++ jamix-business/src/main/java/co/simplon/jamixbusiness/accounts/AccountLoginDto.java
@@ -1,0 +1,11 @@
+package co.simplon.jamixbusiness.accounts;
+
+import jakarta.validation.constraints.Email;
+
+public record AccountLoginDto(@Email(message = "Invalid email format") String email, String password) {
+    @Override
+    public String toString() {
+	return String.format("{email=%s, password=[PROTECTED]}", email);
+    }
+
+}
