+++ jamix-business/src/main/java/co/simplon/jamixbusiness/accounts/AccountRepository.java
@@ -1,0 +1,11 @@
+package co.simplon.jamixbusiness.accounts;
+
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface AccountRepository extends JpaRepository<Account, Long> {
+    Optional<Account> findByEmailIgnoreCase(String email);
+}
