+++ jamix-business/src/main/java/co/simplon/jamixbusiness/accounts/AuthInfo.java
@@ -1,0 +1,8 @@
+package co.simplon.jamixbusiness.accounts;
+
+public record AuthInfo(String token) {
+    @Override
+    public String toString() {
+	return String.format("{token=[PROTECTED]}, accountInfo=%s}");
+    }
+}
