+++ jamix-business/src/main/java/co/simplon/jamixbusiness/accounts/AuthenticationResponse.java
@@ -1,0 +1,18 @@
+package co.simplon.jamixbusiness.accounts;
+
+public class AuthenticationResponse {
+    private String token;
+
+    public AuthenticationResponse(String token) {
+	this.token = token;
+    }
+
+    public String getToken() {
+	return token;
+    }
+
+    public void setToken(String token) {
+	this.token = token;
+    }
+
+}
