--- jamix-business/src/main/java/co/simplon/jamixbusiness/accounts/dtos/AuthInfo.java
@@ -1,8 +1,0 @@
-package co.simplon.jamixbusiness.accounts.dtos;
-
-public record AuthInfo(String token) {
-    @Override
-    public String toString() {
-	return String.format("{token=[PROTECTED]}, accountInfo=%s}");
-    }
-}
