--- jamix-business/src/main/java/co/simplon/jamixbusiness/accounts/dtos/AuthenticationResponse.java
@@ -1,18 +1,0 @@
-package co.simplon.jamixbusiness.accounts.dtos;
-
-public class AuthenticationResponse {
-    private String token;
-
-    public AuthenticationResponse(String token) {
-	this.token = token;
-    }
-
-    public String getToken() {
-	return token;
-    }
-
-    public void setToken(String token) {
-	this.token = token;
-    }
-
-}
