--- jamix-business/src/main/java/co/simplon/jamixbusiness/accounts/dtos/LoginResponse.java
@@ -1,17 +1,0 @@
-package co.simplon.jamixbusiness.accounts.dtos;
-
-public record LoginResponse(String token, String message) {
-    public LoginResponse(String token, String message) {
-	this.token = token;
-	this.message = message;
-    }
-
-    public String token() {
-	return token;
-    }
-
-    public String message() {
-	return message;
-    }
-
-}
