--- jamix-business/src/main/java/co/simplon/jamixbusiness/accounts/repositories/AccountRepository.java
@@ -1,13 +1,0 @@
-package co.simplon.jamixbusiness.accounts.repositories;
-
-import java.util.Optional;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import co.simplon.jamixbusiness.accounts.Account;
-
-@Repository
-public interface AccountRepository extends JpaRepository<Account, Long> {
-    Optional<Account> findByEmailIgnoreCase(String email);
-}
