+++ jamix-business/src/main/java/co/simplon/jamixbusiness/images/ImageService.java
@@ -1,0 +1,15 @@
+package co.simplon.jamixbusiness.images;
+
+import org.springframework.core.io.Resource;
+import org.springframework.web.multipart.MultipartFile;
+
+public interface ImageService {
+    String store(MultipartFile file);
+
+    Resource load(String imageId);
+
+    String getUrl(String imageId);
+
+    void delete(String imageId);
+
+}
