+++ jamix-business/src/main/java/co/simplon/jamixbusiness/locations/LocationService.java
@@ -1,0 +1,14 @@
+package co.simplon.jamixbusiness.locations;
+
+import java.util.List;
+
+public interface LocationService {
+    Boolean isReal(LocationViewDto locationDto);
+
+    Boolean isExist(String city, String zipCode);
+
+    LocationViewDto create(LocationCreateDto dto);
+
+    List<Location> getAllLocation();
+
+}
