+++ jamix-business/src/main/java/co/simplon/jamixbusiness/locations/LocationViewDto.java
@@ -1,0 +1,7 @@
+package co.simplon.jamixbusiness.locations;
+
+import jakarta.validation.constraints.Positive;
+
+public record LocationViewDto(@Positive Long id, String city, String zipCode) {
+
+}
