--- jamix-business/src/main/java/co/simplon/jamixbusiness/locations/dtos/LocationCreateDto.java
@@ -1,12 +1,0 @@
-package co.simplon.jamixbusiness.locations.dtos;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Size;
-
-public record LocationCreateDto(
-
-	@NotBlank @Size(max = 50) String city,
-
-	@NotBlank @Size(max = 5) String zipCode) {
-
-}
