--- jamix-business/src/main/java/co/simplon/jamixbusiness/locations/dtos/LocationViewDto.java
@@ -1,7 +1,0 @@
-package co.simplon.jamixbusiness.locations.dtos;
-
-import jakarta.validation.constraints.Positive;
-
-public record LocationViewDto(@Positive Long id, String city, String zipCode) {
-
-}
