+++ jamix-business/src/main/java/co/simplon/jamixbusiness/offers/dtos/OfferCreateDto.java
@@ -3,22 +3,24 @@ package co.simplon.jamixbusiness.offers.dtos;
 import org.springframework.http.MediaType;
 import org.springframework.web.multipart.MultipartFile;
 
+import co.simplon.jamixbusiness.commons.validations.CleanXSS;
 import co.simplon.jamixbusiness.commons.validations.FileSize;
 import co.simplon.jamixbusiness.commons.validations.FileType;
+import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import jakarta.validation.constraints.Size;
 
-public record OfferCreateDto(@NotBlank @Size(max = 200) String title,
+public record OfferCreateDto(@NotBlank @Size(max = 200) @CleanXSS(message = "Invalid title format") String title,
 
-	@NotBlank @Size(max = 600) String description,
+	@NotBlank @Size(max = 600) @CleanXSS(message = "Invalid description format") String description,
 
-	@Size(max = 255) String contactMail,
+	@Size(max = 255) @Email(message = "Invalid email format") String contactMail,
 
-	@NotBlank @Size(max = 50) String city,
+	@NotBlank @Size(max = 50) @CleanXSS(message = "Invalid city format") String city,
 
-	@NotBlank @Size(max = 5) String zipCode,
+	@NotBlank @Size(max = 5) @CleanXSS(message = "Invalid zip code format") String zipCode,
 
 	@NotNull @Positive Long instrumentId,
