+++ jamix-business/src/main/java/co/simplon/jamixbusiness/offers/services/impl/OfferMusicianServiceImpl.java
@@ -10,12 +10,12 @@ import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.server.ResponseStatusException;
 
 import co.simplon.jamixbusiness.accounts.Account;
-import co.simplon.jamixbusiness.images.services.ImageService;
+import co.simplon.jamixbusiness.images.ImageService;
 import co.simplon.jamixbusiness.locations.Location;
-import co.simplon.jamixbusiness.locations.dtos.LocationCreateDto;
-import co.simplon.jamixbusiness.locations.dtos.LocationViewDto;
-import co.simplon.jamixbusiness.locations.repositories.LocationRepository;
-import co.simplon.jamixbusiness.locations.services.LocationService;
+import co.simplon.jamixbusiness.locations.LocationCreateDto;
+import co.simplon.jamixbusiness.locations.LocationRepository;
+import co.simplon.jamixbusiness.locations.LocationService;
+import co.simplon.jamixbusiness.locations.LocationViewDto;
 import co.simplon.jamixbusiness.offers.Offer;
 import co.simplon.jamixbusiness.offers.dtos.OfferCreateDto;
 import co.simplon.jamixbusiness.offers.dtos.OfferUpdateDto;
@@ -24,11 +24,11 @@ import co.simplon.jamixbusiness.offers.mappers.OfferMapper;
 import co.simplon.jamixbusiness.offers.repositories.OfferRepository;
 import co.simplon.jamixbusiness.offers.services.OfferMusicianService;
 import co.simplon.jamixbusiness.preferences.Goal;
+import co.simplon.jamixbusiness.preferences.GoalRepository;
 import co.simplon.jamixbusiness.preferences.Instrument;
+import co.simplon.jamixbusiness.preferences.InstrumentRepository;
 import co.simplon.jamixbusiness.preferences.Style;
-import co.simplon.jamixbusiness.preferences.repositories.GoalRepository;
-import co.simplon.jamixbusiness.preferences.repositories.InstrumentRepository;
-import co.simplon.jamixbusiness.preferences.repositories.StyleRepository;
+import co.simplon.jamixbusiness.preferences.StyleRepository;
 import co.simplon.jamixbusiness.security.CurrentUserManager;
 
 @Service
