+++ jamix-business/src/main/java/co/simplon/jamixbusiness/preferences/GoalRepository.java
@@ -1,0 +1,11 @@
+package co.simplon.jamixbusiness.preferences;
+
+import java.util.List;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface GoalRepository extends JpaRepository<Goal, Long> {
+
+    List<Goal> findByTypeIn(List<String> goalTypes);
+
+}
