+++ jamix-business/src/main/java/co/simplon/jamixbusiness/preferences/InstrumentRepository.java
@@ -1,0 +1,11 @@
+package co.simplon.jamixbusiness.preferences;
+
+import java.util.List;
+
+import org.springframework.data.domain.Sort;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface InstrumentRepository extends JpaRepository<Instrument, Long> {
+    @Override
+    List<Instrument> findAll(Sort sort);
+}
