+++ jamix-business/src/main/java/co/simplon/jamixbusiness/preferences/PreferenceService.java
@@ -1,0 +1,13 @@
+package co.simplon.jamixbusiness.preferences;
+
+import java.util.List;
+
+public interface PreferenceService {
+
+    List<Instrument> getAllInstruments();
+
+    List<Style> getAllStyle();
+
+    List<Goal> getAllGoal();
+
+}
