--- jamix-business/src/main/java/co/simplon/jamixbusiness/preferences/controllers/PreferenceController.java
@@ -1,39 +1,0 @@
-package co.simplon.jamixbusiness.preferences.controllers;
-
-import java.util.List;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import co.simplon.jamixbusiness.preferences.Goal;
-import co.simplon.jamixbusiness.preferences.Instrument;
-import co.simplon.jamixbusiness.preferences.Style;
-import co.simplon.jamixbusiness.preferences.services.impl.PreferenceServiceImpl;
-
-@RestController
-@RequestMapping("/api")
-public class PreferenceController {
-
-    private final PreferenceServiceImpl service;
-
-    protected PreferenceController(PreferenceServiceImpl service) {
-	this.service = service;
-    }
-
-    @GetMapping("/instruments")
-    public List<Instrument> getAllInstruments() {
-	return service.getAllInstruments();
-    }
-
-    @GetMapping("/styles")
-    public List<Style> getAllStyles() {
-	return service.getAllStyle();
-    }
-
-    @GetMapping("/goals")
-    public List<Goal> getAllGoals() {
-	return service.getAllGoal();
-    }
-
-}
