--- jamix-business/src/main/java/co/simplon/jamixbusiness/preferences/repositories/GoalRepository.java
@@ -1,13 +1,0 @@
-package co.simplon.jamixbusiness.preferences.repositories;
-
-import java.util.List;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import co.simplon.jamixbusiness.preferences.Goal;
-
-public interface GoalRepository extends JpaRepository<Goal, Long> {
-
-    List<Goal> findByTypeIn(List<String> goalTypes);
-
-}
