--- jamix-business/src/main/java/co/simplon/jamixbusiness/preferences/repositories/StyleRepository.java
@@ -1,15 +1,0 @@
-package co.simplon.jamixbusiness.preferences.repositories;
-
-import java.util.List;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import co.simplon.jamixbusiness.preferences.Style;
-
-public interface StyleRepository extends JpaRepository<Style, Long> {
-
-    List<Style> findByNameInOrderByNameAsc(List<String> styleNames);
-
-    List<Style> findByNameIn(List<String> styleNames);
-
-}
