--- jamix-business/src/main/java/co/simplon/jamixbusiness/preferences/services/PreferenceService.java
@@ -1,17 +1,0 @@
-package co.simplon.jamixbusiness.preferences.services;
-
-import java.util.List;
-
-import co.simplon.jamixbusiness.preferences.Goal;
-import co.simplon.jamixbusiness.preferences.Instrument;
-import co.simplon.jamixbusiness.preferences.Style;
-
-public interface PreferenceService {
-
-    List<Instrument> getAllInstruments();
-
-    List<Style> getAllStyle();
-
-    List<Goal> getAllGoal();
-
-}
