+++ jamix-business/src/main/java/co/simplon/jamixbusiness/roles/RoleRepository.java
@@ -1,0 +1,12 @@
+package co.simplon.jamixbusiness.roles;
+
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface RoleRepository extends JpaRepository<Role, Long> {
+    Optional<Role> findByRoleName(String roleName);
+
+}
