--- jamix-business/src/main/java/co/simplon/jamixbusiness/roles/repositories/RoleRepository.java
@@ -1,14 +1,0 @@
-package co.simplon.jamixbusiness.roles.repositories;
-
-import java.util.Optional;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import co.simplon.jamixbusiness.roles.Role;
-
-@Repository
-public interface RoleRepository extends JpaRepository<Role, Long> {
-    Optional<Role> findByRoleName(String roleName);
-
-}
